colourPicker {colourpicker}	R Documentation

Colour picker gadget

Description

This gadget lets you choose colours easily. You can select multiple colours, and you can either choose any RGB colour, or browse through R colours.

Usage

colourPicker(numCols = 3)

Arguments

numCols
The number of colours to select when the gadget launches (you can add and remove more colours from the app itself too)

Value

Vector of selected colours

Note

This gadget returns a vector of colours that can be assigned to a variable. If instead you want to get a text representation of the colours that can embedded into code, use the addin from the RStudio Addins menu.

Examples

if (interactive()) {
  cols <- colourPicker(5)
}

[Package colourpicker version 1.1.0 Index]