colourpicker-package {colourpicker}	R Documentation

colourpicker: A Colour Picker Tool for Shiny and for Selecting Colours in Plots

Description

A colour picker that can be used as an input in 'Shiny' apps or Rmarkdown documents. The colour picker supports alpha opacity, custom colour palettes, and many more options. A Plot Colour Helper tool is available as an 'RStudio' Addin, which helps you pick colours to use in your plots. A more generic Colour Picker 'RStudio' Addin is also provided to let you select colours to use in your R code.

Author(s)

Maintainer: Dean Attali daattali@gmail.com (ORCID)

Other contributors:

David Griswold novachild@gmail.com [contributor]

See Also

Useful links:

https://github.com/daattali/colourpicker

https://daattali.com/shiny/colourInput/

Report bugs at https://github.com/daattali/colourpicker/issues

[Package colourpicker version 1.3.0 Index]